another huge refactor, move some functions to specific elements helpers
