+++ FabricExample/e2e/issue-tests/Test2877.e2e.ts
@@ -1,5 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import { describeIfiOS, selectIssueTestScreen } from '../e2e-utils';
+import { describeIfiOS } from '../e2e-utils';
+import { selectIssueTestScreen } from '../elements/test-screen-navigation';
 
 // issue fixed only on iOS at the moment
 describeIfiOS('Test2877', () => {
