+++ FabricExample/e2e/issue-tests/Test2926.e2e.ts
@@ -1,5 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import { describeIfiOS, selectIssueTestScreen } from '../e2e-utils';
+import { describeIfiOS } from '../e2e-utils';
+import { selectIssueTestScreen } from '../elements/test-screen-navigation';
 import { tapBarBackButton } from '../elements/back-button';
 import { CLASS_NAME_UI_SEARCH_BAR_TEXT_FIELD } from '../native-class-names';
