+++ FabricExample/e2e/issue-tests/Test654.e2e.ts
@@ -1,5 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import { describeIfiOS, selectIssueTestScreen } from '../e2e-utils';
+import { describeIfiOS } from '../e2e-utils';
+import { selectIssueTestScreen } from '../elements/test-screen-navigation';
 import { CLASS_NAME_UI_BUTTON_BAR_BUTTON } from '../native-class-names';
 
 // issue related to iOS native back button
