+++ FabricExample/e2e/single-feature-tests/stack-v5/test-stack-header-selective-updates-ios.e2e.ts
@@ -1,30 +1,13 @@
-import { device, expect, element, by, waitFor } from 'detox';
-import {
-  describeIfiOS,
-  dismissToast,
-  selectPickerOption,
-  selectSingleFeatureTestsScreen,
-} from '../../e2e-utils';
-import {
-  CLASS_NAME_UI_BUTTON_BAR_BUTTON,
-  CLASS_NAME_UI_CONTEXT_MENU_CELL_CONTENT_VIEW,
-} from '../../native-class-names';
+import { device, expect, element, by } from 'detox';
+import { describeIfiOS } from '../../e2e-utils';
+import { selectSingleFeatureTestsScreen } from '../../elements/test-screen-navigation';
+import { selectPickerOption } from '../../elements/settings-controls';
+import { dismissToast } from '../../elements/toast';
+import { checkmarkFor } from '../../elements/context-menu-ios';
+import { CLASS_NAME_UI_BUTTON_BAR_BUTTON } from '../../native-class-names';
 
 function textItem(label: string) {
   return element(by.type(CLASS_NAME_UI_BUTTON_BAR_BUTTON).and(by.label(label)));
-}
-
-// A checked toggle/singleSelection row inside the presented native UIMenu.
-function checkmarkFor(itemLabel: string) {
-  return element(
-    by
-      .id('checkmark')
-      .withAncestor(
-        by
-          .type(CLASS_NAME_UI_CONTEXT_MENU_CELL_CONTENT_VIEW)
-          .and(by.label(itemLabel)),
-      ),
-  );
 }
 
 // `SettingsPicker` derives its option testIDs from the label only, not the item
