+++ FabricExample/e2e/single-feature-tests/stack-v5/test-stack-toolbar-menu-a11y-android.e2e.ts
@@ -1,8 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import {
-  describeIfAndroid,
-  selectSingleFeatureTestsScreen,
-} from '../../e2e-utils';
+import { describeIfAndroid } from '../../e2e-utils';
+import { selectSingleFeatureTestsScreen } from '../../elements/test-screen-navigation';
 
 describeIfAndroid('Stack Toolbar Menu A11y', () => {
   beforeAll(async () => {
