+++ FabricExample/e2e/single-feature-tests/tabs/test-tabs-item-badge.e2e.ts
@@ -1,9 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import {
-  describeIfiOS,
-  isIOSVersionAtLeast,
-  selectSingleFeatureTestsScreen,
-} from '../../e2e-utils';
+import { describeIfiOS, isIOSVersionAtLeast } from '../../e2e-utils';
+import { selectSingleFeatureTestsScreen } from '../../elements/test-screen-navigation';
 import {
   CLASS_NAME_UI_TAB_BAR_BADGE_VIEW_IOS26,
   CLASS_NAME_UI_TAB_BAR_BADGE_VIEW_LEGACY,
