+++ FabricExample/e2e/single-feature-tests/tabs/test-tabs-lifecycle-events.e2e.ts
@@ -1,5 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import { dismissToast, selectSingleFeatureTestsScreen } from '../../e2e-utils';
+import { selectSingleFeatureTestsScreen } from '../../elements/test-screen-navigation';
+import { dismissToast } from '../../elements/toast';
 
 describe('Tabs lifecycle events', () => {
   beforeAll(async () => {
