+++ FabricExample/e2e/single-feature-tests/tabs/test-tabs-prevent-native-selection.e2e.ts
@@ -1,8 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import {
-  selectSingleFeatureTestsScreen,
-  describeIfiPad,
-} from '../../e2e-utils';
+import { describeIfiPad } from '../../e2e-utils';
+import { selectSingleFeatureTestsScreen } from '../../elements/test-screen-navigation';
 import {
   CLASS_NAME_UI_FLOATING_TAB_BAR_ITEM_CELL,
   CLASS_NAME_UI_BUTTON,
