+++ FabricExample/e2e/single-feature-tests/tabs/test-tabs-special-effects-scroll-to-top.e2e.ts
@@ -1,9 +1,7 @@
 import { device, expect, element, by } from 'detox';
-import {
-  scrollUntilVisible,
-  selectSingleFeatureTestsScreen,
-  forceSelectTabByLabel,
-} from '../../e2e-utils';
+import { scrollUntilVisible } from '../../e2e-utils';
+import { selectSingleFeatureTestsScreen } from '../../elements/test-screen-navigation';
+import { forceSelectTabByLabel } from '../../elements/tab-bar';
 import { CLASS_NAME_UI_TAB_BAR } from '../../native-class-names';
 
 describe('Tabs specialEffects — scrollToTop', () => {
