+++ FabricExample/e2e/single-feature-tests/tabs/test-tabs-stale-update-rejection.e2e.ts
@@ -1,5 +1,6 @@
 import { device, expect, element, by } from 'detox';
-import { dismissToast, selectSingleFeatureTestsScreen } from '../../e2e-utils';
+import { selectSingleFeatureTestsScreen } from '../../elements/test-screen-navigation';
+import { dismissToast } from '../../elements/toast';
 
 // These scenarios are split into two separate suites using `device.reloadReactNative()`.
 // Running these scenarios sequentially within a single app lifecycle caused deterministic
