+++ apps/src/tests/single-feature-tests/stack-v5/test-stack-back-button-android/scenario-description.ts
@@ -2,7 +2,7 @@ import type { ScenarioDescription } from '@apps/tests/shared/helpers';
 
 export const scenarioDescription: ScenarioDescription = {
   name: 'Stack Back Button',
-  key: 'test-stack-back-button',
+  key: 'test-stack-back-button-android',
   details: 'Tests back button customization: hidden, tint color, custom icon.',
   platforms: ['android'],
   e2eCoverage: 'incomplete',
